make 2cd slot a boolean for installp.
The 2cd slot used to be a list of computers to limit package installation
to those computers only. But this limited the slot to exactly 1 case.

Now it's replaced with a simple boolean, that can be used for any use
case. The boolean logic is embedded into the list itself.