revamp the alternative to key-chord-mode.
It incorrectly tried to dynamically bind the function name`message'
with a let. In a lisp-2 the let only works on variable names, not
function names. The message was actually turned off by just commenting
it out, not by dynamically overriding it in a let.

Also put the lambda in a named variable for readability.